Rename scvi docker image to sc_tools and add map_cell_types task
