+++ docker/sc_tools/build.env
@@ -1,0 +1,10 @@
+# Tool versions
+SC_TOOLS_VERSION=1.0.0
+PYTHON3_VERSION=3.10.12
+JAX_VERSION=0.4.28
+CUDA_VERSION=12.3.0
+CELL_TYPE_MAPPER_VERSION=1.5.3
+
+# Image info
+IMAGE_NAME=sc_tools
+IMAGE_TAG=${SC_TOOLS_VERSION}
