+++ docker/sc_tools/resources/s_genes_tirosh_hm.txt
