+++ docker/sc_tools/scripts/README.md
