+++ docker/sc_tools/scripts/main/DEPRECATED_annotate_cells.py
