+++ docker/sc_tools/scripts/main/add_harmony.py
