+++ docker/sc_tools/scripts/main/artifact_metrics.py
