+++ docker/sc_tools/scripts/main/empty
