+++ docker/sc_tools/scripts/main/filter.py
