+++ docker/sc_tools/scripts/main/merge_and_plot_qc.py
