+++ docker/sc_tools/scripts/main/mmc.py
@@ -1,0 +1,99 @@
+from cell_type_mapper.cli.map_to_on_the_fly_markers import OnTheFlyMapper
+from pathlib import Path
+import os
+import argparse
+
+
+# Works best on CPU
+# os.environ["AIBS_BKP_USE_TORCH"] = "false"
+# os.environ["NUMEXPR_NUM_THREADS"] = "1"
+# os.environ["MKL_NUM_THREADS"] = "1"
+# os.environ["OMP_NUM_THREADS"] = "1"
+
+# These were some initial defaults
+CHUNK_SIZE = 40000
+N_RUNNERS_UP = 5
+RNG_SEED = 11235813
+# N_PROCESSORS = 8
+# MAX_GB = 48.0
+
+
+def main(args: argparse.Namespace):
+    EXTENDED_RESULTS_JSON = f"{args.output_prefix}.extended_results.json"
+    CSV_RESULTS = f"{args.output_prefix}.results.csv"
+    LOG_FILE = f"{args.output_prefix}.log.txt"
+
+    print(f"EXTENDED_RESULTS: {EXTENDED_RESULTS_JSON}")
+    print(f"CSV_RESULTS: {CSV_RESULTS}")
+    print(f"LOG_FILE: {LOG_FILE}")
+
+    # MMC
+    config = {
+        "precomputed_stats": {
+            "path": args.mmc_taxonomy_path
+        },
+        "type_assignment": {
+            "bootstrap_factor": 0.5,
+            "bootstrap_iteration": 100,
+            "normalization": "raw",
+            "rng_seed": RNG_SEED,
+            "chunk_size": CHUNK_SIZE,
+            "n_runners_up": N_RUNNERS_UP
+        },
+        "reference_markers": {
+            "log2_fold_min_th": 0.5
+        },
+        "query_markers": {
+            "n_per_utility": 15,
+            "genes_at_a_time": 1
+        },
+        "query_path": args.adata_input,
+        "extended_result_path": EXTENDED_RESULTS_JSON,
+        "csv_result_path": CSV_RESULTS,
+        "log_path": LOG_FILE,
+        "cloud_safe": True,
+        "verbose_csv": True,
+        "map_to_ensembl": True,
+    }
+
+    # Creates the results files: .json, .csv, .txt
+    # https://github.com/AllenInstitute/cell_type_mapper/blob/main/docs/output.md
+    runner = OnTheFlyMapper(args=[], input_data=config)
+    runner.run()
+
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser(description="Mape cell types with Allen Institute's Cell Type Mapper")
+    parser.add_argument(
+        "--adata-input",
+        type=str,
+        required=True,
+        help="AnnData object for a dataset",
+    )
+    parser.add_argument(
+        "--mmc-taxonomy-path",
+        type=str,
+        required=True,
+        help="Path to MMC taxonomy files",
+    )
+    parser.add_argument(
+        "--output-prefix",
+        type=str,
+        required=True,
+        help="Prefix used to name output files",
+    )
+    args = parser.parse_args()
+    main(args)
+
+
+# # SEA-AD "references"
+# MTG
+# DLPFC
+# MTG_DLPFC
+
+# look at discrepancies.
+
+# ## xylena reference...
+# xylena_train for Taxonomy
+
+# test on xylena_test + xylena_query
