+++ docker/sc_tools/scripts/main/plot_feats_and_groups.py
