+++ docker/sc_tools/scripts/main/prep_metadata.py
