+++ docker/sc_tools/scripts/utility/empty
