+++ docker/sc_tools/scripts/utility/helpers.py
