--- docker/scvi/build.env
@@ -1,11 +1,0 @@
-# Tool versions
-SCVI_VERSION=1.2.0
-GCLOUD_CLI_VERSION=444.0.0
-PYTHON3_VERSION=3.10.12
-JAX_VERSION=0.4.28
-CUDA_VERSION=12.3.0
-REVISION=1
-
-# Image info
-IMAGE_NAME=scvi
-IMAGE_TAG=${SCVI_VERSION}_${REVISION}
