+++ workflows/cohort_analysis/cluster_data/cluster_data.wdl
@@ -91,7 +91,7 @@ task integrate_sample_data {
 	}
 
 	runtime {
-		docker: "~{container_registry}/scvi:1.2.0_1"
+		docker: "~{container_registry}/sc_tools:1.0.0"
 		cpu: 2
 		memory: "~{mem_gb} GB"
 		disks: "local-disk ~{disk_size} HDD"
@@ -132,7 +132,7 @@ task cluster_cells {
 	}
 
 	runtime {
-		docker: "~{container_registry}/scvi:1.2.0_1"
+		docker: "~{container_registry}/sc_tools:1.0.0"
 		cpu: 16
 		memory: "~{mem_gb} GB"
 		disks: "local-disk ~{disk_size} HDD"
