+++ workflows/preprocess/preprocess.wdl
@@ -423,7 +423,7 @@ task counts_to_adata {
 	}
 
 	runtime {
-		docker: "~{container_registry}/scvi:1.2.0_1"
+		docker: "~{container_registry}/sc_tools:1.0.0"
 		cpu: 2
 		memory: "16 GB"
 		disks: "local-disk ~{disk_size} HDD"
